Adv 2

+++ 0x0C-more_malloc_free/101-mul.c
@@ -11,7 +11,7 @@
  */
 int main(int argc, char **argv)
 {
-	int num1, num2;
+	unsigned long num1, num2;
 
 	if (argc != 3)
 	{
